+++ conf/urls.py
@@ -240,6 +240,16 @@ urlpatterns = [
         name='export-plan-target-market-documents-list-create',
     ),
     url(
+        r'^exportplan/funding-credit-options/(?P<pk>[0-9]+)/$',
+        exportplan.views.FundingCreditOptionsUpdateDestroyView.as_view(),
+        name='export-plan-funding-credit-options-detail-update',
+    ),
+    url(
+        r'^exportplan/funding-credit-options/$',
+        exportplan.views.FundingCreditOptionsCreateAPIView.as_view(),
+        name='export-plan-funding-credit-options-list-create',
+    ),
+    url(
         r'^dataservices/easeofdoingbusiness/(?P<country_code>.*)/$',
         dataservices.views.RetrieveEaseOfBusinessIndex.as_view(),
         name='dataservices-easeofdoingbusiness-index',
@@ -273,6 +283,11 @@ urlpatterns = [
         r'^dataservices/cia-factbook-data/$',
         dataservices.views.RetrieveCiaFactbooklDataView.as_view(),
         name='cia-factbook-data',
+    ),
+    url(
+        r'^dataservices/society-data-by-country/$',
+        dataservices.views.RetrieveSocietyDataByCountryView.as_view(),
+        name='dataservices-society-data-by-country',
     ),
     url(
         r'^dataservices/historicalimportdata/$',