+++ dataservices/helpers.py
@@ -52,8 +52,11 @@ class ComTradeData:
             else:
                 return ''
 
-    def get_url(self):
-        url_options = f'&r={self.reporting_area_id}&p={self.partner_country_id}&cc={self.product_code}&ps=All&rg=1'
+    def get_url(self, from_uk=False):
+        if from_uk:
+            url_options = f'&r={self.partner_country_id}&p={self.reporting_area_id}&cc={self.product_code}&ps=All&rg=2'
+        else:
+            url_options = f'&r={self.reporting_area_id}&p=0&cc={self.product_code}&ps=All&rg=1'
         return self.url + url_options
 
     def get_product_code(self, commodity_code):
@@ -61,12 +64,7 @@ class ComTradeData:
         return ''.join(commodity_list)
 
     def get_last_year_import_data(self, from_uk=False):
-
-        url = self.get_url()
-        if from_uk:
-            url_options = f'&r={self.reporting_area_id}&p={self.partner_country_id}&cc={self.product_code}&ps=All&rg=2'
-            url = self.url + url_options
-
+        url = self.get_url(from_uk=from_uk)
         comdata = requests.get(url)
         if comdata and 'dataset' in comdata.json() and comdata.json()['dataset']:
             comdata_df = pandas.DataFrame.from_dict(comdata.json()['dataset']).sort_values(by='period', ascending=False)
@@ -77,7 +75,6 @@ class ComTradeData:
                 year_on_year_change = None
                 # check if data available for more than one year
                 if len(comdata_df.index) > 1:
-
                     try:
                         last_year_import = comdata_df[comdata_df.period == comdata_df.period.max() - 1][
                             'TradeValue'
@@ -367,27 +364,48 @@ def get_cia_factbook_data(country_name, data_keys=None):
         return {}
 
 
+@TTLCache()
 def get_internet_usage(country):
     try:
         internet_usage_obj = models.InternetUsage.objects.filter(country_name=country).latest()
-    except models.InternetUsage.DoesNotExist:
-        return {}
-    return {
-        'internet_usage': {
-            'value': '{:.2f}'.format(internet_usage_obj.value) if hasattr(internet_usage_obj, 'value') else None,
-            'year': internet_usage_obj.year if hasattr(internet_usage_obj, 'year') else None,
+        return {
+            'internet_usage': {
+                'value': '{:.2f}'.format(internet_usage_obj.value) if hasattr(internet_usage_obj, 'value') else None,
+                'year': internet_usage_obj.year if hasattr(internet_usage_obj, 'year') else None,
+            }
         }
-    }
+    except Exception:
+        return {}
 
 
+@TTLCache()
 def get_cpi_data(country):
     try:
         cpi_obj = models.ConsumerPriceIndex.objects.filter(country_name=country).latest()
-    except models.ConsumerPriceIndex.DoesNotExist:
+        return {
+            'cpi': {
+                'value': '{:.2f}'.format(cpi_obj.value) if hasattr(cpi_obj, 'value') else None,
+                'year': cpi_obj.year,
+            }
+        }
+    except Exception:
         return {}
-    return {
-        'cpi': {'value': '{:.2f}'.format(cpi_obj.value) if hasattr(cpi_obj, 'value') else None, 'year': cpi_obj.year}
-    }
+
+
+@TTLCache()
+def get_society_data(country):
+    society_data = {}
+    cia_people_data = get_cia_factbook_data(country, data_keys=['people'])
+
+    if not cia_people_data:
+        return society_data
+
+    cia_people_data = cia_people_data.get('people')
+
+    society_data['religions'] = cia_people_data.get('religions', {})
+    society_data['languages'] = cia_people_data.get('languages', {})
+
+    return society_data
 
 
 def millify(n):
@@ -421,3 +439,12 @@ def get_serialized_instance_from_model(model_class, serializer_class, filter_arg
         return serializer.data
     except model_class.DoesNotExist:
         return None
+
+
+def calculate_total_internet_population(internet_usage, total_population):
+    if internet_usage and total_population:
+        percent = float(internet_usage.get('value')) / 100
+        total = total_population.get('total_population', 0) * 1000
+        return millify(percent * total)
+    else:
+        return ''