+++ dataservices/tests/test_views.py
@@ -68,6 +68,12 @@ def country_data():
 
 
 @pytest.fixture(autouse=True)
+def society_data():
+    country = models.Country.objects.create(iso2='UK', name='United Kingdom')
+    models.RuleOfLaw.objects.create(iso2='UK', country_name='United Kingdom', rank=10, score=76, country=country)
+
+
+@pytest.fixture(autouse=True)
 def cia_factbook_data():
     return factories.CIAFactBookFactory()
 
@@ -78,7 +84,15 @@ def test_get_easeofdoingbusiness(api_client):
 
     response = api_client.get(url)
     assert response.status_code == 200
-    assert response.json() == {'country_name': 'China', 'country_code': 'CN', 'year_2019': 10, 'total': 2}
+    assert response.json() == {
+        'country_name': 'China',
+        'country_code': 'CN',
+        'year_2019': 10,
+        'total': 2,
+        'country': None,
+        'year': '2019',
+        'rank': 10,
+    }
 
 
 @pytest.mark.django_db
@@ -96,7 +110,15 @@ def test_get_corruptionperceptionsindex(api_client):
 
     response = api_client.get(url)
     assert response.status_code == 200
-    assert response.json() == {'country_name': 'China', 'country_code': 'CN', 'cpi_score_2019': 10, 'rank': 3}
+    assert response.json() == {
+        'country_name': 'China',
+        'country_code': 'CN',
+        'cpi_score_2019': 10,
+        'rank': 3,
+        'country': None,
+        'total': 2,
+        'year': '2019',
+    }
 
 
 @pytest.mark.django_db
@@ -124,6 +146,7 @@ def test_get_world_economic_outlook(api_client):
             'units': 'dollars',
             'year_2020': '21234141.000',
             'year_2021': '32432423.000',
+            'country': None,
         },
         {
             'country_code': 'CN',
@@ -133,6 +156,7 @@ def test_get_world_economic_outlook(api_client):
             'units': 'Percent change',
             'year_2020': '323.210',
             'year_2021': '1231.100',
+            'country': None,
         },
     ]
 
@@ -225,14 +249,28 @@ def test_get_country_data(api_client):
 
     response = api_client.get(url)
     assert response.status_code == 200
-
     assert response.json() == {
         'country_data': {
-            'consumer_price_index': {'country_name': 'Canada', 'country_code': 'CNN', 'value': '20.560', 'year': 2019},
-            'internet_usage': {'country_name': 'Canada', 'country_code': 'CNN', 'value': '20.230', 'year': 2019},
+            'consumer_price_index': {
+                'country_name': 'Canada',
+                'country_code': 'CNN',
+                'value': '20.560',
+                'year': 2019,
+                'country': None,
+            },
+            'internet_usage': {
+                'country_name': 'Canada',
+                'country_code': 'CNN',
+                'value': '20.230',
+                'year': 2019,
+                'country': None,
+                'total_internet_usage': '7.70 million',
+            },
             'corruption_perceptions_index': None,
             'ease_of_doing_bussiness': None,
             'gdp_per_capita': None,
+            'income': None,
+            'total_population': '38.07 million',
         }
     }
 
@@ -243,7 +281,6 @@ def test_get_country_data_not_found(api_client):
 
     response = api_client.get(url)
     assert response.status_code == 200
-
     assert response.json() == {
         'country_data': {
             'consumer_price_index': None,
@@ -251,6 +288,8 @@ def test_get_country_data_not_found(api_client):
             'corruption_perceptions_index': None,
             'ease_of_doing_bussiness': None,
             'gdp_per_capita': None,
+            'income': None,
+            'total_population': '0.00',
         }
     }
 
@@ -266,10 +305,19 @@ def test_get_country_data_cpi_not_found(api_client):
     assert response.json() == {
         'country_data': {
             'consumer_price_index': None,
-            'internet_usage': {'country_name': 'Canada', 'country_code': 'CNN', 'value': '20.230', 'year': 2019},
+            'internet_usage': {
+                'country_name': 'Canada',
+                'country_code': 'CNN',
+                'value': '20.230',
+                'year': 2019,
+                'total_internet_usage': '7.70 million',
+                'country': None,
+            },
             'corruption_perceptions_index': None,
             'ease_of_doing_bussiness': None,
             'gdp_per_capita': None,
+            'income': None,
+            'total_population': '38.07 million',
         },
     }
 
@@ -281,14 +329,21 @@ def test_get_country_data_internet_not_found(api_client):
 
     response = api_client.get(url)
     assert response.status_code == 200
-
     assert response.json() == {
         'country_data': {
-            'consumer_price_index': {'country_name': 'Canada', 'country_code': 'CNN', 'value': '20.560', 'year': 2019},
+            'consumer_price_index': {
+                'country_name': 'Canada',
+                'country_code': 'CNN',
+                'value': '20.560',
+                'year': 2019,
+                'country': None,
+            },
             'internet_usage': None,
             'corruption_perceptions_index': None,
             'ease_of_doing_bussiness': None,
             'gdp_per_capita': None,
+            'income': None,
+            'total_population': '38.07 million',
         }
     }
 
@@ -391,6 +446,7 @@ def test_population_data_by_country(api_client, internet_usage_data):
             'urban_population_total': 56970,
             'urban_population_percentage_formatted': '83.53% (56.97 million)',
             'total_population': '68.20 million',
+            'total_population_raw': 68204000,
             'cpi': {'value': '150.56', 'year': 2019},
         }
     ]
@@ -413,6 +469,7 @@ def test_population_data_by_country_multiple_countries(api_client, internet_usag
             'urban_population_total': 56970,
             'urban_population_percentage_formatted': '83.53% (56.97 million)',
             'total_population': '68.20 million',
+            'total_population_raw': 68204000,
             'cpi': {'value': '150.56', 'year': 2019},
         },
         {
@@ -423,6 +480,7 @@ def test_population_data_by_country_multiple_countries(api_client, internet_usag
             'urban_population_total': 64044,
             'urban_population_percentage_formatted': '76.33% (64.04 million)',
             'total_population': '83.90 million',
+            'total_population_raw': 83902000,
         },
     ]
 
@@ -452,3 +510,107 @@ def test_suggested_countries_api_without_hs_code(client):
     assert response.status_code == 500
     json_dict = json.loads(response.content)
     assert json_dict['error_message'] == "hs_code missing in request params"
+
+
+@pytest.mark.django_db
+def test_income_data_api(api_client):
+    # import countries and income data
+    management.call_command('import_countries')
+    management.call_command('import_income_data')
+
+    url = reverse('dataservices-country-data', kwargs={'country': 'Canada'})
+    json_response = api_client.get(url).json()
+    assert 'income' in json_response['country_data']
+    assert 'Canada' == json_response['country_data']['income']['country_name']
+    assert '37653.281' == json_response['country_data']['income']['value']
+    # Retrieve India too as it has cpi data mocked
+    url = reverse('dataservices-country-data', kwargs={'country': 'India'})
+    json_response = api_client.get(url).json()
+    assert 'income' in json_response['country_data']
+    assert 'India' == json_response['country_data']['income']['country_name']
+    assert '1735.329' == json_response['country_data']['income']['value']
+    assert 5 == json_response['country_data']['ease_of_doing_bussiness']['year_2019']
+    assert 2 == json_response['country_data']['ease_of_doing_bussiness']['total']
+    assert '2019' == json_response['country_data']['ease_of_doing_bussiness']['year']
+    assert 9 == json_response['country_data']['corruption_perceptions_index']['rank']
+    assert 2 == json_response['country_data']['corruption_perceptions_index']['total']
+    assert '2019' == json_response['country_data']['corruption_perceptions_index']['year']
+
+
+@pytest.mark.django_db
+def test_society_data_by_country_with_country_arg_missing(api_client):
+    url = reverse('dataservices-society-data-by-country')
+
+    response = api_client.get(url)
+
+    assert response.status_code == 400
+
+
+@pytest.mark.django_db
+def test_society_data_by_country_with_country_not_found(api_client):
+    url = reverse('dataservices-society-data-by-country')
+    response = api_client.get(url, data={'countries': 'abcde'})
+
+    assert response.status_code == 200
+
+    assert response.json() == [{'country': 'abcde', 'rule_of_law': None}]
+
+
+@pytest.mark.django_db
+def test_society_data_by_country(api_client):
+    url = reverse('dataservices-society-data-by-country')
+
+    response = api_client.get(url, data={'countries': 'United Kingdom'})
+
+    assert response.status_code == 200
+    assert response.json() == [
+        {
+            'country': 'United Kingdom',
+            'languages': {
+                'date': '2012',
+                'note': 'test data',
+                'language': [
+                    {
+                        'name': 'English',
+                    }
+                ],
+            },
+            'religions': {
+                "date": "2011",
+                'religion': [
+                    {
+                        'name': 'Christian',
+                        'note': 'includes Anglican, Roman Catholic, Presbyterian, Methodist',
+                        'percent': 59.5,
+                    },
+                    {'name': 'Muslim', 'percent': 4.4},
+                    {'name': 'Hindu', 'percent': 1.3},
+                    {'name': 'other', 'percent': 2},
+                    {'name': 'unspecified', 'percent': 7.2},
+                    {'name': 'none', "percent": 25.7},
+                ],
+            },
+            'rule_of_law': {
+                'country_name': 'United Kingdom',
+                'iso2': 'UK',
+                'rank': 10,
+                'score': '76.000',
+            },
+        }
+    ]
+
+
+@pytest.mark.django_db
+def test_society_data_repr():
+    rule_of_law = models.RuleOfLaw.objects.create(iso2='CN', country_name='Canada', rank=10, score=76)
+
+    assert str(rule_of_law) == 'Canada'
+
+
+@pytest.mark.django_db
+def test_currencies_data_repr():
+    currencies = models.Currency.objects.create(
+        iso2='IN', country_name='India', currency_name='Indian Rupee', alphabetic_code='INR', numeric_code=123
+    )
+
+    assert str(currencies) == 'India'